use CapsuleCastAll, adjust interactable items. Change camera distance. Adjust finish points.
Update Food UIText
Todos:
The interact is not sensitive,  smaller the capsule.
Player can go on the table.
The food and amenity game, the items are too small, suggest zoom in after entering the room.
The guest line looks strange, because the order need to follow but it is parallel line.
Make the signature board move with camera, so don't need to adjust after camera change.